Move serval method to interface
