+++ bcos-rpc/bcos-rpc/jsonrpc/JsonRpcImpl_2_0.h
@@ -117,157 +117,15 @@ public:
     void getGroupBlockNumber(RespFunc _respFunc) override;
 
 public:
-    void callI(const Json::Value& req, RespFunc _respFunc)
-    {
-        call(req[0u].asString(), req[1u].asString(), req[2u].asString(), req[3u].asString(),
-            _respFunc);
-    }
-
-    void sendTransactionI(const Json::Value& req, RespFunc _respFunc)
-    {
-        sendTransaction(req[0u].asString(), req[1u].asString(), req[2u].asString(),
-            req[3u].asBool(), _respFunc);
-    }
-
-    void getTransactionI(const Json::Value& req, RespFunc _respFunc)
-    {
-        getTransaction(req[0u].asString(), req[1u].asString(), req[2u].asString(), req[3u].asBool(),
-            _respFunc);
-    }
-
-    void getTransactionReceiptI(const Json::Value& req, RespFunc _respFunc)
-    {
-        getTransactionReceipt(req[0u].asString(), req[1u].asString(), req[2u].asString(),
-            req[3u].asBool(), _respFunc);
-    }
-
-    void getBlockByHashI(const Json::Value& req, RespFunc _respFunc)
-    {
-        getBlockByHash(req[0u].asString(), req[1u].asString(), req[2u].asString(),
-            (req.size() > 3 ? req[3u].asBool() : true), (req.size() > 4 ? req[4u].asBool() : true),
-            _respFunc);
-    }
-
-    void getBlockByNumberI(const Json::Value& req, RespFunc _respFunc)
-    {
-        getBlockByNumber(req[0u].asString(), req[1u].asString(), req[2u].asInt64(),
-            (req.size() > 3 ? req[3u].asBool() : true), (req.size() > 4 ? req[4u].asBool() : true),
-            _respFunc);
-    }
-
-    void getBlockHashByNumberI(const Json::Value& req, RespFunc _respFunc)
-    {
-        getBlockHashByNumber(req[0u].asString(), req[1u].asString(), req[2u].asInt64(), _respFunc);
-    }
-
-    void getBlockNumberI(const Json::Value& req, RespFunc _respFunc)
-    {
-        boost::ignore_unused(req);
-        getBlockNumber(req[0u].asString(), req[1u].asString(), _respFunc);
-    }
-
-    void getCodeI(const Json::Value& req, RespFunc _respFunc)
-    {
-        getCode(req[0u].asString(), req[1u].asString(), req[2u].asString(), _respFunc);
-    }
-
-    void getABII(const Json::Value& req, RespFunc _respFunc)
-    {
-        getABI(req[0u].asString(), req[1u].asString(), req[2u].asString(), _respFunc);
-    }
-
-    void getSealerListI(const Json::Value& req, RespFunc _respFunc)
-    {
-        boost::ignore_unused(req);
-        getSealerList(req[0u].asString(), req[1u].asString(), _respFunc);
-    }
-
-    void getObserverListI(const Json::Value& req, RespFunc _respFunc)
-    {
-        boost::ignore_unused(req);
-        getObserverList(req[0u].asString(), req[1u].asString(), _respFunc);
-    }
-
-    void getPbftViewI(const Json::Value& req, RespFunc _respFunc)
-    {
-        boost::ignore_unused(req);
-        getPbftView(req[0u].asString(), req[1u].asString(), _respFunc);
-    }
-
-    void getPendingTxSizeI(const Json::Value& req, RespFunc _respFunc)
-    {
-        boost::ignore_unused(req);
-        getPendingTxSize(req[0u].asString(), req[1u].asString(), _respFunc);
-    }
-
-    void getSyncStatusI(const Json::Value& req, RespFunc _respFunc)
-    {
-        boost::ignore_unused(req);
-        getSyncStatus(req[0u].asString(), req[1u].asString(), _respFunc);
-    }
-
-    void getConsensusStatusI(const Json::Value& _req, RespFunc _respFunc)
-    {
-        getConsensusStatus(_req[0u].asString(), _req[1u].asString(), _respFunc);
-    }
-
-    void getSystemConfigByKeyI(const Json::Value& req, RespFunc _respFunc)
-    {
-        getSystemConfigByKey(req[0u].asString(), req[1u].asString(), req[2u].asString(), _respFunc);
-    }
-
-    void getTotalTransactionCountI(const Json::Value& req, RespFunc _respFunc)
-    {
-        boost::ignore_unused(req);
-        getTotalTransactionCount(req[0u].asString(), req[1u].asString(), _respFunc);
-    }
-
-    void getPeersI(const Json::Value& req, RespFunc _respFunc)
-    {
-        boost::ignore_unused(req);
-        getPeers(_respFunc);
-    }
-
-    void getGroupPeersI(const Json::Value& req, RespFunc _respFunc)
-    {
-        getGroupPeers(req[0u].asString(), _respFunc);
-    }
-
-    // get all the groupID list
-    void getGroupListI(const Json::Value& _req, RespFunc _respFunc)
-    {
-        (void)_req;
-        getGroupList(_respFunc);
-    }
-    // get the group information of the given group
-    void getGroupInfoI(const Json::Value& _req, RespFunc _respFunc)
-    {
-        (void)_req;
-        getGroupInfo(_req[0u].asString(), _respFunc);
-    }
-    // get the group information of the given group
-    void getGroupInfoListI(const Json::Value& _req, RespFunc _respFunc)
-    {
-        (void)_req;
-        getGroupInfoList(_respFunc);
-    }
-    // get the information of a given node
-    void getGroupNodeInfoI(const Json::Value& _req, RespFunc _respFunc)
-    {
-        getGroupNodeInfo(_req[0u].asString(), _req[1u].asString(), _respFunc);
-    }
-
     void setNodeInfo(const NodeInfo& _nodeInfo) { m_nodeInfo = _nodeInfo; }
     NodeInfo nodeInfo() const { return m_nodeInfo; }
     GroupManager::Ptr groupManager() { return m_groupManager; }
 
 protected:
-    void initMethod();
     static std::shared_ptr<bcos::bytes> decodeData(const std::string& _data);
-    static void parseRpcRequestJson(const std::string& _requestBody, JsonRequest& _jsonRequest);
+    
     static void parseRpcResponseJson(const std::string& _responseBody, JsonResponse& _jsonResponse);
-    static Json::Value toJsonResponse(const JsonResponse& _jsonResponse);
-    static std::string toStringResponse(const JsonResponse& _jsonResponse);
+    
     static void toJsonResp(
         Json::Value& jResp, bcos::protocol::Transaction::ConstPtr _transactionPtr);
 
@@ -279,11 +137,8 @@ protected:
     static void addProofToResponse(
         Json::Value& jResp, std::string const& _key, ledger::MerkleProofPtr _merkleProofPtr);
 
-    void onRPCRequest(const std::string& _requestBody, Sender _sender) override;
-
     virtual void handleRpcRequest(std::shared_ptr<boostssl::MessageFace> _msg,
         std::shared_ptr<boostssl::ws::WsSession> _session);
-
 
     // TODO: check perf influence
     NodeService::Ptr getNodeService(
@@ -308,9 +163,6 @@ private:
     void getGroupPeers(std::string const& _groupID, RespFunc _respFunc) override;
 
 private:
-    std::unordered_map<std::string, std::function<void(Json::Value, RespFunc _respFunc)>>
-        m_methodToFunc;
-
     GroupManager::Ptr m_groupManager;
     bcos::gateway::GatewayInterface::Ptr m_gatewayInterface;
     std::shared_ptr<boostssl::ws::WsService> m_wsService;
